requirements.txt
pyTelegramBotAPI==4.14.0
Flask==2.3.3
 HEAD
Pillow==10.3.0

Pillow==10.3.0
 5d776aea62a13e35e0aa578dec1455e0e69966b6
